backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
